refactor: pipeline api consumption into hook, comment out unused imports(for now), ignore tsbuildinfo, bump version
