+++ frontend/src/App.tsx
@@ -22,36 +22,11 @@ import { Button } from '@/components/ui/button';
 import { LoadingSpinner } from '@/components/ui/loading-spinner';
 import { Link } from 'react-router-dom';
 import useBackendStatus from './hooks/use-backend';
-import { useState } from 'react';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from './components/ui/card';
+import usePipeline from './hooks/use-pipeline';
 
 function App() {
-	const [userInput, setUserInput] = useState('');
-	const [loading, setLoading] = useState(false);
-	const [response, setResponse] = useState(null);
-
-	const handleInputChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-		setUserInput(event.target.value);
-	};
-
-	const handleSubmit = async () => {
-		setLoading(true);
-		try {
-			const response = await fetch('http://localhost:3000/api/runPipeline', {
-				method: 'POST',
-				headers: {
-					'Content-Type': 'application/json',
-				},
-				body: JSON.stringify({ userInput: userInput }),
-			});
-			const data = await response.json();
-			setResponse(data);
-		} catch (error) {
-			console.error('Error:', error);
-		} finally {
-			setLoading(false);
-		}
-	};
+	const { userInput, loading, response, handleInputChange, handleSubmit } = usePipeline();
 	return (
 		<ThemeProvider defaultTheme="system" storageKey="vite-ui-theme">
 			<div className="flex flex-col h-full">
@@ -115,11 +90,11 @@ function App() {
 						<Card>
 							<CardHeader>
 								<CardTitle>Time</CardTitle>
-								<CardDescription>Estimated time to complete the task via RAG pipeline</CardDescription>
+								<CardDescription>Estimated time required to complete the task</CardDescription>
 							</CardHeader>
 							<CardContent>
 								<Button variant="destructive">
-									<pre>{JSON.stringify(response.result, null, 2).replace(/\"/g, '')}</pre>
+									<pre>{response}</pre>
 								</Button>
 							</CardContent>
 						</Card>
